v.0.3.0: Moving types to the types.ts file
